csv-python/jcsv/__init__.py

import jpype
import jpype.imports
import jcsv.jvm


__csv = jpype.JPackage("io.github.gciatto.csv")


Table = __csv.Table
Row = __csv.Row
Record = __csv.Record
Header = __csv.Header
Formatter = __csv.Formatter
Parser = __csv.Parser
Csv = __csv.Csv
